+++ src/frontend/js/widgets/Dashboard/components/DashboardItem/Order/DashboardItemOrder.spec.tsx
@@ -37,7 +37,7 @@ import { expectBannerError } from 'utils/test/expectBanner';
 import { expectNoSpinner, expectSpinner } from 'utils/test/expectSpinner';
 import { Deferred } from 'utils/test/deferred';
 import { expectBreadcrumbsToEqualParts } from 'utils/test/expectBreadcrumbsToEqualParts';
-import { mockProductWithOrder } from 'utils/test/mockProductWithOrder';
+import { mockCourseProductWithOrder } from 'utils/test/mockCourseProductWithOrder';
 import { LearnerDashboardPaths } from '../../../utils/learnerRouteMessages';
 import { DashboardTest } from '../../DashboardTest';
 import { DashboardItemOrder } from './DashboardItemOrder';
@@ -108,7 +108,7 @@ describe('<DashboardItemOrder/>', () => {
   it('renders a pending order', async () => {
     const order: Order = OrderFactory({ state: OrderState.PENDING }).one();
     order.target_courses = [];
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} />, { wrapper });
 
@@ -121,7 +121,7 @@ describe('<DashboardItemOrder/>', () => {
   it('renders an order with certificate', async () => {
     const order: Order = OrderFactory({ certificate: faker.string.uuid() }).one();
     order.target_courses = [];
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     const certificate: Certificate = {
       ...CertificateFactory({
@@ -151,7 +151,7 @@ describe('<DashboardItemOrder/>', () => {
   it('does not render an order with certificate', async () => {
     const order: Order = OrderFactory({ certificate: faker.string.uuid() }).one();
     order.target_courses = [];
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} />, { wrapper });
 
@@ -169,7 +169,7 @@ describe('<DashboardItemOrder/>', () => {
   it('renders a non-writable order without target courses without certificate', async () => {
     const order: Order = OrderFactory().one();
     order.target_courses = [];
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} />, { wrapper });
 
@@ -181,7 +181,7 @@ describe('<DashboardItemOrder/>', () => {
 
   it('renders a non-writable order with target courses', async () => {
     const order: Order = OrderFactory().one();
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} />, { wrapper });
 
@@ -205,7 +205,7 @@ describe('<DashboardItemOrder/>', () => {
 
     order.enrollments[0].course_run.state.priority = Priority.ONGOING_OPEN;
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} />, { wrapper });
 
@@ -233,7 +233,7 @@ describe('<DashboardItemOrder/>', () => {
       enrollments: [],
     }).one();
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} />, { wrapper });
 
@@ -254,7 +254,7 @@ describe('<DashboardItemOrder/>', () => {
   it('renders a writable order with no target courses', async () => {
     const order: Order = OrderFactory().one();
     order.target_courses = [];
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} writable={true} showDetailsButton={false} />, {
       wrapper,
@@ -276,7 +276,7 @@ describe('<DashboardItemOrder/>', () => {
       },
     ];
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} writable={true} showDetailsButton={false} />, {
       wrapper,
@@ -318,7 +318,7 @@ describe('<DashboardItemOrder/>', () => {
       target_courses: TargetCourseFactory().many(1),
       enrollments: [],
     }).one();
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
     fetchMock.post('https://joanie.endpoint/api/v1.0/enrollments/', []);
     fetchMock.get(
       'https://joanie.endpoint/api/v1.0/orders/',
@@ -397,7 +397,7 @@ describe('<DashboardItemOrder/>', () => {
       enrollments: [],
     }).one();
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
     fetchMock.get(
       'https://joanie.endpoint/api/v1.0/orders/',
       { results: [order], next: null, previous: null, count: null },
@@ -452,7 +452,7 @@ describe('<DashboardItemOrder/>', () => {
       [],
     );
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
     fetchMock.post('https://joanie.endpoint/api/v1.0/enrollments/', []);
     fetchMock.get(
       'https://joanie.endpoint/api/v1.0/orders/',
@@ -549,7 +549,7 @@ describe('<DashboardItemOrder/>', () => {
       },
     ];
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
     fetchMock.post('https://joanie.endpoint/api/v1.0/enrollments/', []);
     fetchMock.get(
       'https://joanie.endpoint/api/v1.0/orders/',
@@ -625,7 +625,7 @@ describe('<DashboardItemOrder/>', () => {
     // When the existing enrollment will be set as is_active: true.
     fetchMock.put('https://joanie.endpoint/api/v1.0/enrollments/' + enrollment.id + '/', []);
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
     fetchMock.get(
       'https://joanie.endpoint/api/v1.0/orders/',
       { results: [order], next: null, previous: null, count: null },
@@ -706,7 +706,7 @@ describe('<DashboardItemOrder/>', () => {
         }).many(1),
       }).many(1),
     }).one();
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} writable={true} showDetailsButton={false} />, {
       wrapper,
@@ -742,7 +742,7 @@ describe('<DashboardItemOrder/>', () => {
     // Make target course enrolled.
     order.enrollments = EnrollmentFactory({ course_run: courseRun }).many(1);
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} writable={true} showDetailsButton={false} />, {
       wrapper,
@@ -772,7 +772,7 @@ describe('<DashboardItemOrder/>', () => {
       enrollments: [],
     }).one();
 
-    const product = mockProductWithOrder(order);
+    const { product } = mockCourseProductWithOrder(order);
 
     render(<DashboardItemOrder order={order} writable={true} showDetailsButton={false} />, {
       wrapper,